pyproject.toml
[project]
name = "endofactory"
version = "0.1.1"
description = "Revolutionary EndoVQA dataset construction tool for rapid dataset mixing and configuration"
authors = [
    {name = "TiramisuQiao",email = "tlmsq@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "polars>=0.20.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "uuid>=1.30",
    "pytest (>=8.4.1,<9.0.0)",
]
license = {text = "MIT"}
keywords = ["endoscopy", "medical", "VQA", "dataset", "polars", "yaml", "cli"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Scientific/Engineering :: Information Analysis",
]

[project.urls]
Homepage = "https://github.com/your-org/EndoFactory"
Repository = "https://github.com/your-org/EndoFactory"
Issues = "https://github.com/your-org/EndoFactory/issues"

[tool.poetry]
packages = [{include = "endofactory", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0"
pytest-benchmark = ">=4.0"

[project.scripts]
endofactory = "endofactory.cli:app"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-q"
markers = [
  "performance: marks performance tests (deselect with '-m \"not performance\"')",
]

[project.optional-dependencies]
tests = [
  "pytest>=7.0",
  "pytest-benchmark>=4.0",
]
